Made ADC Clock have a 512 S/H Time

+++ trunk/UoC Drum/FuncLib/Peripherals/ADC/adc12.h
@@ -14,7 +14,7 @@ void ADC12_Init(void);
 void ADC12_SetStartAddress(uint16_t startAdd);
 uint8_t ADC12_GetStartAddress(void);
 void ADC12_SetupAddress(uint8_t channel, uint8_t settings);
-void ADC12_SH_Time(uint8_t SampleHold);
+void ADC12_SH_Time(uint16_t SampleHold);
 
 /* Sample and return on selected StartAddress's Pin*/
 uint16_t ADC12_Sample(void);
